scrapers/autotraders/requirements.txt
async-generator==1.10
attrs==22.1.0
boto3==1.26.18
botocore==1.29.18
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
jmespath==1.0.1
numpy==1.23.5
outcome==1.2.0
pandas==1.5.2
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
selenium==4.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.7
urllib3==1.26.13
websockets==10.4
wsproto==1.2.0